Solve stupid bugs and change matrix operation res mem to calloc()

+++ src/utils/Activations/relu/relu.c
@@ -4,7 +4,7 @@ dARRAY * forward_pass_relu(dARRAY * layer_matrix){
   dARRAY * relu_out = (dARRAY*)malloc(sizeof(dARRAY));
   relu_out->matrix = (double*)malloc(sizeof(double)*layer_matrix->shape[0]*layer_matrix->shape[1]);
   for(int i=0;i<layer_matrix->shape[0]*layer_matrix->shape[1];i++)
-    relu_out->matrix[i] = layer_matrix->matrix[i]>(double)0 ? layer_matrix->matrix[i] : 0;
+    relu_out->matrix[i] = layer_matrix->matrix[i]>(double)0 ?(double)layer_matrix->matrix[i] : 0.0;
   relu_out->shape[0] = layer_matrix->shape[0];
   relu_out->shape[1] = layer_matrix->shape[1];
   return relu_out;
@@ -14,7 +14,7 @@ dARRAY * backward_pass_relu(dARRAY * layer_matrix){
   dARRAY * relu_out = (dARRAY*)malloc(sizeof(dARRAY));
   relu_out->matrix = (double*)malloc(sizeof(double)*layer_matrix->shape[0]*layer_matrix->shape[1]);
   for(int i=0;i<layer_matrix->shape[0]*layer_matrix->shape[1];i++)
-    relu_out->matrix[i] = layer_matrix->matrix[i]>(double)0 ? 1 : 0;
+    relu_out->matrix[i] = layer_matrix->matrix[i]>(double)0 ? 1.0 : 0.0;
   relu_out->shape[0] = layer_matrix->shape[0];
   relu_out->shape[1] = layer_matrix->shape[1];
   return relu_out;
